Make root_uri a proper string?
