+++ test/test_workspace.py
@@ -3,6 +3,7 @@ import os
 import sys
 import os.path as osp
 from pyls import uris
+from pyls._utils import to_text_string
 
 PY2 = sys.version_info.major == 2
 
@@ -17,28 +18,6 @@ DOC_URI = uris.from_fs_path(__file__)
 
 def path_as_uri(path):
     return pathlib.Path(osp.abspath(path)).as_uri()
-
-
-def to_text_string(obj, encoding=None):
-    """Convert `obj` to (unicode) text string"""
-    if PY2:
-        if isinstance(obj, unicode):
-            return obj
-        # Python 2
-        if encoding is None:
-            return unicode(obj)
-        else:
-            return unicode(obj, encoding)
-    else:
-        # Python 3
-        if encoding is None:
-            return str(obj)
-        elif isinstance(obj, str):
-            # In case this function is not used properly, this could happen
-            return obj
-        else:
-            return str(obj, encoding)
-
 
 def test_local(pyls):
     """ Since the workspace points to the test directory """
